Refactor TestES256K test into new jwx_es256k_test.go file
The TestES256K method has been moved from the crypto/jwx_test.go to a new jwx_es256k_test.go file that is triggerd by the jwx_es256k flag.
